+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 2.x branch
 ## 2.21 branch
 ### 2.21.3
-* added `checkbox()` and `select()` method for `BootstrapFormHelper`;
+* added `checkbox()`, `radio()` and `select()` methods for `BootstrapFormHelper`;
 * fixed templates for `BootstrapFormHelper` with input groups;
 * fixed little bug for `BootstrapFormHelper::submit()` method.